apserv-51-aiohttp-simple/web/requirements.txt
ulid-py
aiohttp
asyncio
async_timeout
attrs
chardet
cchardet
multidict
yarl
aiodns
hypercorn